+++ backend/core/src/listings/dto/listing.dto.ts
@@ -23,7 +23,7 @@ import { AddressCreateDto, AddressDto, AddressUpdateDto } from "../../shared/dto
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { UserBasicDto } from "../../auth/dto/user.dto"
 import { ListingStatus } from "../types/listing-status-enum"
-import { ListingFilterKeys } from "../types/listing-filter-keys-enum"
+import { AvailabilityFilterEnum, ListingFilterKeys } from "../types/listing-filter-keys-enum"
 import { PaginationFactory, PaginationAllowsAllQueryParams } from "../../shared/dto/pagination.dto"
 import { BaseFilter } from "../../shared/dto/filter.dto"
 import { UnitCreateDto, UnitDto, UnitUpdateDto } from "../../units/dto/unit.dto"
@@ -795,6 +795,14 @@ export class ListingFilterParams extends BaseFilter {
 
   @Expose()
   @ApiProperty({
+    enum: Object.keys(AvailabilityFilterEnum),
+    example: "hasAvailability",
+    required: false,
+  })
+  [ListingFilterKeys.availability]?: AvailabilityFilterEnum;
+
+  @Expose()
+  @ApiProperty({
     type: Boolean,
     example: "true",
     required: false,
@@ -848,12 +856,21 @@ export class ListingsRetrieveQueryParams {
   view?: string
 }
 
+const FilterKeysList = { ...ListingFilterKeys, ...AvailabilityFilterEnum }
+type FilterKeysList = typeof FilterKeysList
+
 // Using a record lets us enforce that all types are handled in addFilter
-export const filterTypeToFieldMap: Record<keyof typeof ListingFilterKeys, string> = {
+export const filterTypeToFieldMap: Record<keyof typeof FilterKeysList, string> = {
   status: "listings.status",
   name: "listings.name",
   neighborhood: "property.neighborhood",
   bedrooms: "unitTypeRef.num_bedrooms",
   zipcode: "buildingAddress.zipCode",
   seniorHousing: "reservedCommunityType.name",
+  // Fields for the availability are determined based on the value of the filter, not the
+  // key. Keep this bogus value to prevent the filter from being rejected.
+  availability: "",
+  hasAvailability: "unitsSummary.total_available",
+  noAvailability: "unitsSummary.total_available",
+  waitlist: "listings.is_waitlist_open",
 }