+++ backend/core/src/listings/listings.service.ts
@@ -42,6 +42,7 @@ export class ListingsService {
       .leftJoin("property.units", "units")
       .leftJoin("units.unitType", "unitTypeRef")
       .leftJoin("property.buildingAddress", "buildingAddress")
+      .leftJoin("listings.unitsSummary", "unitsSummary")
       .leftJoin("listings.reservedCommunityType", "reservedCommunityType")
       .groupBy("listings.id")
       .orderBy({ "listings.id": "DESC" })