+++ backend/core/src/shared/filter/custom_filters.ts
@@ -1,6 +1,10 @@
 import { WhereExpression } from "typeorm"
-import { ListingFilterKeys } from "../../listings/types/listing-filter-keys-enum"
+import {
+  AvailabilityFilterEnum,
+  ListingFilterKeys,
+} from "../../listings/types/listing-filter-keys-enum"
 import { filterTypeToFieldMap } from "../../listings/dto/listing.dto"
+import { Compare } from "../dto/filter.dto"
 
 export function addSeniorHousingQuery(qb: WhereExpression, filterValue: string) {
   const whereParameterName = ListingFilterKeys.seniorHousing
@@ -19,5 +23,34 @@ export function addSeniorHousingQuery(qb: WhereExpression, filterValue: string) 
         [whereParameterName]: seniorHousingCommunityType,
       }
     )
+  }
+}
+
+function addAvailabilityParams(
+  qb: WhereExpression,
+  filterType: AvailabilityFilterEnum,
+  comparison: string,
+  filterValue: any
+) {
+  const hasAvailabilityColumnName = `LOWER(CAST(${filterTypeToFieldMap[filterType]} as text))`
+  const whereParameterName = filterType
+  qb.andWhere(`${hasAvailabilityColumnName} ${comparison} LOWER(:${whereParameterName})`, {
+    [whereParameterName]: filterValue,
+  })
+}
+
+export function addAvailabilityQuery(qb: WhereExpression, filterValue: AvailabilityFilterEnum) {
+  switch (filterValue) {
+    case AvailabilityFilterEnum.hasAvailability:
+      addAvailabilityParams(qb, filterValue, Compare[">="], 1)
+      return
+    case AvailabilityFilterEnum.noAvailability:
+      addAvailabilityParams(qb, filterValue, Compare["="], 0)
+      return
+    case AvailabilityFilterEnum.waitlist:
+      addAvailabilityParams(qb, filterValue, Compare["="], true)
+      return
+    default:
+      return
   }
 }