+++ backend/core/src/shared/filter/index.ts
@@ -1,8 +1,11 @@
 import { HttpException, HttpStatus } from "@nestjs/common"
 import { WhereExpression } from "typeorm"
 import { Compare } from "../dto/filter.dto"
-import { ListingFilterKeys } from "../../listings/types/listing-filter-keys-enum"
-import { addSeniorHousingQuery } from "./custom_filters"
+import {
+  AvailabilityFilterEnum,
+  ListingFilterKeys,
+} from "../../listings/types/listing-filter-keys-enum"
+import { addSeniorHousingQuery, addAvailabilityQuery } from "./custom_filters"
 
 /**
  *
@@ -70,6 +73,10 @@ export function addFilters<FilterParams, FilterFieldMap>(
           // Handle custom filters here, before dropping into generic filter handler
           if (filterType == ListingFilterKeys.seniorHousing) {
             addSeniorHousingQuery(qb, filterValue)
+            return
+          }
+          if (filterType == ListingFilterKeys.availability) {
+            addAvailabilityQuery(qb, filterValue as AvailabilityFilterEnum)
             return
           }