+++ backend/core/types/src/backend-swagger.ts
@@ -3715,7 +3715,10 @@ export interface ListingFilterParams {
   /**  */
   zipcode?: string
 
-  /** */
+  /**  */
+  availability?: EnumListingFilterParamsAvailability
+
+  /**  */
   seniorHousing?: boolean
 }
 
@@ -5725,6 +5728,11 @@ export enum EnumListingFilterParamsStatus {
   "active" = "active",
   "pending" = "pending",
   "closed" = "closed",
+}
+export enum EnumListingFilterParamsAvailability {
+  "hasAvailability" = "hasAvailability",
+  "noAvailability" = "noAvailability",
+  "waitlist" = "waitlist",
 }
 export enum ListingApplicationAddressType {
   "leasingAgent" = "leasingAgent",